requirements.txt
numpy
Pillow
pygame
tqdm
matplotlib
opencv-python==4.5.1.48

yt_dlp
vidgear

gym-retro
gym==0.21.0

english_words